connected server to client
