+++ Backend/utils/config.ts
@@ -1,0 +1,10 @@
+import dotenv from "dotenv";
+dotenv.config();
+
+const PORT = process.env.PORT;
+
+if (!PORT) {
+  throw new Error("secrets failed");
+}
+
+export default { PORT };
