+++ Frontend/lib/services/game/service.ts
@@ -1,0 +1,1 @@
+const baseUrl: string = "/api/games";
